OTT bookmarks - prevent stop twice on destroy (#16)
Because of the event dispatching order, destroy is handled by the plugin before stop. Then, in the next time stop is sent it contains an incorrect position.

The workaround is that if destroy has sent the stop event, stop will not send it again.

This behavior needs to be revisited if/when there's a change in event dispatching order.